requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==26.0.0
faker_food==0.3.0
furl==2.1.3
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
openapi-codec==1.3.2
orderedmultidict==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
setuptools==72.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0
